requirements.txt
pytest
requests
ddt
pyyaml
selenium
webdriver-manager
